Refactoring tests into multiple files; moving away from macros
