+++ ava.config.js
@@ -1,5 +1,5 @@
 export default {
-  files: ['dist/test/*.spec.js'],
+  files: ['dist/test/**/*.spec.js'],
   failWithoutAssertions: false,
   environmentVariables: {
     ts_node_install_lock: `id-${Math.floor(Math.random() * 10e9)}`,
