+++ package.json
@@ -133,7 +133,7 @@
     "semver": "^7.1.3",
     "throat": "^6.0.1",
     "typedoc": "^0.20.28",
-    "typescript": "4.3.5",
+    "typescript": "4.4.3",
     "typescript-json-schema": "^0.42.0",
     "util.promisify": "^1.0.1"
   },
