+++ src/test/index.spec.ts
@@ -1,33 +1,14 @@
 import { test } from './testlib';
 import { expect } from 'chai';
-import * as exp from 'expect';
-import {
-  ChildProcess,
-  exec as childProcessExec,
-  ExecException,
-  ExecOptions,
-} from 'child_process';
-import { dirname, join, resolve, sep as pathSep } from 'path';
-import { homedir, tmpdir } from 'os';
+import { join, resolve, sep as pathSep } from 'path';
+import { tmpdir } from 'os';
 import semver = require('semver');
 import ts = require('typescript');
-import type * as tsNodeTypes from '../index';
-import * as fs from 'fs';
-import { unlinkSync, existsSync, lstatSync, mkdtempSync } from 'fs';
-import { NodeFS, npath } from '@yarnpkg/fslib';
-import * as promisify from 'util.promisify';
-import { sync as rimrafSync } from 'rimraf';
+import { lstatSync, mkdtempSync } from 'fs';
+import { npath } from '@yarnpkg/fslib';
 import type _createRequire from 'create-require';
-const createRequire: typeof _createRequire = require('create-require');
 import { pathToFileURL } from 'url';
-import type * as Module from 'module';
-import { PassThrough } from 'stream';
-import * as getStream from 'get-stream';
-import { once } from 'lodash';
-import { upstreamTopLevelAwaitTests } from './node-repl-tla';
-import { createMacrosAndHelpers, ExecMacroAssertionCallback } from './macros';
-
-const xfs = new NodeFS(fs);
+import { createExec } from './exec-helpers';
 
 async function settled<T>(fn: () => Promise<T> | T) {
   try {
@@ -43,61 +24,6 @@ async function settled<T>(fn: () => Promise<T> | T) {
   }
 }
 
-interface CreateReplViaApiOptions {
-  createReplOpts?: Partial<tsNodeTypes.CreateReplOptions>;
-  createServiceOpts?: Partial<tsNodeTypes.CreateOptions>;
-}
-function createReplViaApi({
-  createReplOpts,
-  createServiceOpts,
-}: CreateReplViaApiOptions = {}) {
-  const stdin = new PassThrough();
-  const stdout = new PassThrough();
-  const stderr = new PassThrough();
-  const replService = createRepl({
-    stdin,
-    stdout,
-    stderr,
-    ...createReplOpts,
-  });
-  const service = create({
-    ...replService.evalAwarePartialHost,
-    project: `${TEST_DIR}/tsconfig.json`,
-    ...createServiceOpts,
-  });
-  replService.setService(service);
-  return { stdin, stdout, stderr, replService, service };
-}
-
-// Todo combine with replApiMacro
-async function executeInRepl(
-  input: string,
-  {
-    waitMs = 1e3,
-    startOptions,
-    ...rest
-  }: CreateReplViaApiOptions & {
-    waitMs?: number;
-    startOptions?: Parameters<typeof replService.startInternal>[0];
-  } = {}
-) {
-  const { stdin, stdout, stderr, replService } = createReplViaApi(rest);
-
-  replService.startInternal(startOptions);
-
-  stdin.write(input);
-  stdin.end();
-  await promisify(setTimeout)(waitMs);
-  stdout.end();
-  stderr.end();
-
-  return {
-    stdin,
-    stdout: await getStream(stdout),
-    stderr: await getStream(stderr),
-  };
-}
-
 import {
   BIN_CWD_PATH,
   BIN_PATH,
@@ -108,18 +34,19 @@ import {
   TEST_DIR,
   installTsNode,
   testsDirRequire,
-} from './before-all';
+  tsNodeTypes,
+  xfs,
+} from './helpers';
 
-const { exec, createExecMacro } = createMacrosAndHelpers({
-  test,
-  defaultCwd: TEST_DIR,
+const exec = createExec({
+  cwd: TEST_DIR,
 });
 
 // Set after ts-node is installed locally
-let { create, VERSION, createRepl }: typeof tsNodeTypes = {} as any;
+let { create, VERSION }: typeof tsNodeTypes = {} as any;
 test.beforeAll(async () => {
   await installTsNode();
-  ({ create, VERSION, createRepl } = testsDirRequire('ts-node'));
+  ({ create, VERSION } = testsDirRequire('ts-node'));
 });
 
 test.suite('ts-node', (test) => {
@@ -433,629 +360,6 @@ test.suite('ts-node', (test) => {
       expect(err).to.equal(null);
       expect(stdout).to.equal('undefined\n');
     });
-
-    test('should run REPL when --interactive passed and stdin is not a TTY', async () => {
-      const execPromise = exec(`${cmd} --interactive`);
-      execPromise.child.stdin!.end('console.log("123")\n');
-      const { err, stdout } = await execPromise;
-      expect(err).to.equal(null);
-      expect(stdout).to.equal('> 123\n' + 'undefined\n' + '> ');
-    });
-
-    test('REPL has command to get type information', async () => {
-      const execPromise = exec(`${cmd} --interactive`);
-      execPromise.child.stdin!.end('\nconst a = 123\n.type a');
-      const { err, stdout } = await execPromise;
-      expect(err).to.equal(null);
-      expect(stdout).to.equal(
-        '> undefined\n' + '> undefined\n' + '> const a: 123\n' + '> '
-      );
-    });
-
-    const execMacro = createExecMacro({
-      cmd,
-      cwd: TEST_DIR,
-    });
-
-    type ReplApiMacroAssertions = (
-      stdout: string,
-      stderr: string
-    ) => Promise<void>;
-
-    const replApiMacro = test.macro(
-      (opts: { input: string }, assertions: ReplApiMacroAssertions) => async (
-        t
-      ) => {
-        const { input } = opts;
-        const { stdin, stdout, stderr, replService } = createReplViaApi();
-        replService.start();
-        stdin.write(input);
-        stdin.end();
-        await promisify(setTimeout)(1e3);
-        stdout.end();
-        stderr.end();
-        const stderrString = await getStream(stderr);
-        const stdoutString = await getStream(stdout);
-        await assertions(stdoutString, stderrString);
-      }
-    );
-
-    // Serial because it's timing-sensitive
-    test.serial(
-      'REPL can be created via API',
-      replApiMacro,
-      {
-        input: '\nconst a = 123\n.type a\n',
-      },
-      async (stdout, stderr) => {
-        expect(stderr).to.equal('');
-        expect(stdout).to.equal(
-          '> undefined\n' + '> undefined\n' + '> const a: 123\n' + '> '
-        );
-      }
-    );
-
-    // Serial because it's timing-sensitive
-    test.serial('REPL can be configured on `start`', async () => {
-      const prompt = '#> ';
-
-      const { stdout, stderr } = await executeInRepl('const x = 3', {
-        startOptions: {
-          prompt,
-          ignoreUndefined: true,
-        },
-      });
-
-      expect(stderr).to.equal('');
-      expect(stdout).to.equal(`${prompt}${prompt}`);
-    });
-
-    // Serial because it's timing-sensitive
-    test.serial(
-      'REPL uses a different context when `useGlobal` is false',
-      async () => {
-        const { stdout, stderr } = await executeInRepl(
-          // No error when re-declaring x
-          'const x = 3\n' +
-            // console.log ouput will end up in the stream and not in test output
-            'console.log(1)\n',
-          {
-            startOptions: {
-              useGlobal: false,
-            },
-          }
-        );
-
-        expect(stderr).to.equal('');
-        expect(stdout).to.equal(`> undefined\n> 1\nundefined\n> `);
-      }
-    );
-
-    test.suite(
-      '[eval], <repl>, and [stdin] execute with correct globals',
-      (test) => {
-        interface GlobalInRepl extends NodeJS.Global {
-          testReport: any;
-          replReport: any;
-          stdinReport: any;
-          evalReport: any;
-          module: any;
-          exports: any;
-          fs: any;
-          __filename: any;
-          __dirname: any;
-        }
-        const globalInRepl = global as GlobalInRepl;
-        const programmaticTest = test.macro(
-          (
-            {
-              evalCodeBefore,
-              stdinCode,
-            }: {
-              evalCodeBefore: string | null;
-              stdinCode: string;
-            },
-            assertions: (stdout: string) => Promise<void> | void
-          ) => async (t) => {
-            delete globalInRepl.testReport;
-            delete globalInRepl.replReport;
-            delete globalInRepl.stdinReport;
-            delete globalInRepl.evalReport;
-            delete globalInRepl.module;
-            delete globalInRepl.exports;
-            delete globalInRepl.fs;
-            delete globalInRepl.__filename;
-            delete globalInRepl.__dirname;
-            const { stdin, stderr, stdout, replService } = createReplViaApi();
-            if (typeof evalCodeBefore === 'string') {
-              replService.evalCode(evalCodeBefore);
-            }
-            replService.start();
-            stdin.write(stdinCode);
-            stdin.end();
-            await promisify(setTimeout)(1e3);
-            stdout.end();
-            stderr.end();
-            expect(await getStream(stderr)).to.equal('');
-            await assertions(await getStream(stdout));
-          }
-        );
-
-        const declareGlobals = `declare var replReport: any, stdinReport: any, evalReport: any, restReport: any, global: any, __filename: any, __dirname: any, module: any, exports: any, fs: any;`;
-        function setReportGlobal(type: 'repl' | 'stdin' | 'eval') {
-          return `
-            ${declareGlobals}
-            global.${type}Report = {
-              __filename: typeof __filename !== 'undefined' && __filename,
-              __dirname: typeof __dirname !== 'undefined' && __dirname,
-              moduleId: typeof module !== 'undefined' && module.id,
-              modulePath: typeof module !== 'undefined' && module.path,
-              moduleFilename: typeof module !== 'undefined' && module.filename,
-              modulePaths: typeof module !== 'undefined' && [...module.paths],
-              exportsTest: typeof exports !== 'undefined' ? module.exports === exports : null,
-              stackTest: new Error().stack!.split('\\n')[1],
-              moduleAccessorsTest: typeof fs === 'undefined' ? null : fs === require('fs'),
-              argv: [...process.argv]
-            };
-          `.replace(/\n/g, '');
-        }
-        const reportsObject = `
-          {
-            stdinReport: typeof stdinReport !== 'undefined' && stdinReport,
-            evalReport: typeof evalReport !== 'undefined' && evalReport,
-            replReport: typeof replReport !== 'undefined' && replReport
-          }
-        `;
-        const printReports = `
-          ${declareGlobals}
-          console.log(JSON.stringify(${reportsObject}));
-        `.replace(/\n/g, '');
-        const saveReportsAsGlobal = `
-          ${declareGlobals}
-          global.testReport = ${reportsObject};
-        `.replace(/\n/g, '');
-
-        function parseStdoutStripReplPrompt(stdout: string) {
-          // Strip node's welcome header, only uncomment if running these tests manually against vanilla node
-          // stdout = stdout.replace(/^Welcome to.*\nType "\.help" .*\n/, '');
-          expect(stdout.slice(0, 2)).to.equal('> ');
-          expect(stdout.slice(-12)).to.equal('undefined\n> ');
-          return parseStdout(stdout.slice(2, -12));
-        }
-        function parseStdout(stdout: string) {
-          return JSON.parse(stdout);
-        }
-
-        /** Every possible ./node_modules directory ascending upwards starting with ./tests/node_modules */
-        const modulePaths = createModulePaths(TEST_DIR);
-        const rootModulePaths = createModulePaths(ROOT_DIR);
-        function createModulePaths(dir: string) {
-          const modulePaths: string[] = [];
-          for (let path = dir; ; path = dirname(path)) {
-            modulePaths.push(join(path, 'node_modules'));
-            if (dirname(path) === path) break;
-          }
-          return modulePaths;
-        }
-
-        // Executable is `ts-node` on Posix, `bin.js` on Windows due to Windows shimming limitations (this is determined by package manager)
-        const tsNodeExe = exp.stringMatching(/\b(ts-node|bin.js)$/);
-
-        test(
-          'stdin',
-          execMacro,
-          {
-            stdin: `${setReportGlobal('stdin')};${printReports}`,
-            flags: '',
-          },
-          (stdout) => {
-            const report = parseStdout(stdout);
-            exp(report).toMatchObject({
-              stdinReport: {
-                __filename: '[stdin]',
-                __dirname: '.',
-                moduleId: '[stdin]',
-                modulePath: '.',
-                // Note: vanilla node does does not have file extension
-                moduleFilename: join(TEST_DIR, `[stdin].ts`),
-                modulePaths,
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(TEST_DIR, `[stdin].ts`)}:1:`
-                ),
-                moduleAccessorsTest: null,
-                argv: [tsNodeExe],
-              },
-              evalReport: false,
-              replReport: false,
-            });
-          }
-        );
-        test(
-          'repl',
-          execMacro,
-          {
-            stdin: `${setReportGlobal('repl')};${printReports}`,
-            flags: '-i',
-          },
-          (stdout) => {
-            const report = parseStdoutStripReplPrompt(stdout);
-            exp(report).toMatchObject({
-              stdinReport: false,
-              evalReport: false,
-              replReport: {
-                __filename: false,
-                __dirname: false,
-                moduleId: '<repl>',
-                modulePath: '.',
-                moduleFilename: null,
-                modulePaths: exp.objectContaining({
-                  ...[join(TEST_DIR, `repl/node_modules`), ...modulePaths],
-                }),
-                // Note: vanilla node REPL does not set exports
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(TEST_DIR, '<repl>.ts')}:2:`
-                ),
-                moduleAccessorsTest: true,
-                argv: [tsNodeExe],
-              },
-            });
-            // Prior to these, nyc adds another entry on Windows; we need to ignore it
-            exp(report.replReport.modulePaths.slice(-3)).toMatchObject([
-              join(homedir(), `.node_modules`),
-              join(homedir(), `.node_libraries`),
-              // additional entry goes to node's install path
-              exp.any(String),
-            ]);
-          }
-        );
-
-        // Should ignore -i and run the entrypoint
-        test(
-          '-i w/entrypoint ignores -i',
-          execMacro,
-          {
-            stdin: `${setReportGlobal('repl')};${printReports}`,
-            flags: '-i ./repl/script.js',
-          },
-          (stdout) => {
-            const report = parseStdout(stdout);
-            exp(report).toMatchObject({
-              stdinReport: false,
-              evalReport: false,
-              replReport: false,
-            });
-          }
-        );
-
-        // Should not execute stdin
-        // Should not interpret positional arg as an entrypoint script
-        test(
-          '-e',
-          execMacro,
-          {
-            stdin: `throw new Error()`,
-            flags: `-e "${setReportGlobal('eval')};${printReports}"`,
-          },
-          (stdout) => {
-            const report = parseStdout(stdout);
-            exp(report).toMatchObject({
-              stdinReport: false,
-              evalReport: {
-                __filename: '[eval]',
-                __dirname: '.',
-                moduleId: '[eval]',
-                modulePath: '.',
-                // Note: vanilla node does does not have file extension
-                moduleFilename: join(TEST_DIR, `[eval].ts`),
-                modulePaths: [...modulePaths],
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(TEST_DIR, `[eval].ts`)}:1:`
-                ),
-                moduleAccessorsTest: true,
-                argv: [tsNodeExe],
-              },
-              replReport: false,
-            });
-          }
-        );
-        test(
-          '-e w/entrypoint arg does not execute entrypoint',
-          execMacro,
-          {
-            stdin: `throw new Error()`,
-            flags: `-e "${setReportGlobal(
-              'eval'
-            )};${printReports}" ./repl/script.js`,
-          },
-          (stdout) => {
-            const report = parseStdout(stdout);
-            exp(report).toMatchObject({
-              stdinReport: false,
-              evalReport: {
-                __filename: '[eval]',
-                __dirname: '.',
-                moduleId: '[eval]',
-                modulePath: '.',
-                // Note: vanilla node does does not have file extension
-                moduleFilename: join(TEST_DIR, `[eval].ts`),
-                modulePaths,
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(TEST_DIR, `[eval].ts`)}:1:`
-                ),
-                moduleAccessorsTest: true,
-                argv: [tsNodeExe, './repl/script.js'],
-              },
-              replReport: false,
-            });
-          }
-        );
-        test(
-          '-e w/non-path arg',
-          execMacro,
-          {
-            stdin: `throw new Error()`,
-            flags: `-e "${setReportGlobal(
-              'eval'
-            )};${printReports}" ./does-not-exist.js`,
-          },
-          (stdout) => {
-            const report = parseStdout(stdout);
-            exp(report).toMatchObject({
-              stdinReport: false,
-              evalReport: {
-                __filename: '[eval]',
-                __dirname: '.',
-                moduleId: '[eval]',
-                modulePath: '.',
-                // Note: vanilla node does does not have file extension
-                moduleFilename: join(TEST_DIR, `[eval].ts`),
-                modulePaths,
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(TEST_DIR, `[eval].ts`)}:1:`
-                ),
-                moduleAccessorsTest: true,
-                argv: [tsNodeExe, './does-not-exist.js'],
-              },
-              replReport: false,
-            });
-          }
-        );
-        test(
-          '-e -i',
-          execMacro,
-          {
-            stdin: `${setReportGlobal('repl')};${printReports}`,
-            flags: `-e "${setReportGlobal('eval')}" -i`,
-          },
-          (stdout) => {
-            const report = parseStdoutStripReplPrompt(stdout);
-            exp(report).toMatchObject({
-              stdinReport: false,
-              evalReport: {
-                __filename: '[eval]',
-                __dirname: '.',
-                moduleId: '[eval]',
-                modulePath: '.',
-                // Note: vanilla node does does not have file extension
-                moduleFilename: join(TEST_DIR, `[eval].ts`),
-                modulePaths,
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(TEST_DIR, `[eval].ts`)}:1:`
-                ),
-                moduleAccessorsTest: true,
-                argv: [tsNodeExe],
-              },
-              replReport: {
-                __filename: '[eval]',
-                __dirname: '.',
-                moduleId: '<repl>',
-                modulePath: '.',
-                moduleFilename: null,
-                modulePaths: exp.objectContaining({
-                  ...[join(TEST_DIR, `repl/node_modules`), ...modulePaths],
-                }),
-                // Note: vanilla node REPL does not set exports, so this would be false
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(TEST_DIR, '<repl>.ts')}:2:`
-                ),
-                moduleAccessorsTest: true,
-                argv: [tsNodeExe],
-              },
-            });
-            // Prior to these, nyc adds another entry on Windows; we need to ignore it
-            exp(report.replReport.modulePaths.slice(-3)).toMatchObject([
-              join(homedir(), `.node_modules`),
-              join(homedir(), `.node_libraries`),
-              // additional entry goes to node's install path
-              exp.any(String),
-            ]);
-          }
-        );
-
-        test(
-          '-e -i w/entrypoint ignores -e and -i, runs entrypoint',
-          execMacro,
-          {
-            stdin: `throw new Error()`,
-            flags: '-e "throw new Error()" -i ./repl/script.js',
-          },
-          (stdout) => {
-            const report = parseStdout(stdout);
-            exp(report).toMatchObject({
-              stdinReport: false,
-              evalReport: false,
-              replReport: false,
-            });
-          }
-        );
-
-        test(
-          '-e -i when -e throws error, -i does not run',
-          execMacro,
-          {
-            stdin: `console.log('hello')`,
-            flags: `-e "throw new Error('error from -e')" -i`,
-            expectError: true,
-          },
-          (stdout, stderr, err) => {
-            exp(err).toBeDefined();
-            exp(stdout).toBe('');
-            exp(stderr).toContain('error from -e');
-          }
-        );
-
-        // Serial because it's timing-sensitive
-        test.serial(
-          'programmatically, eval-ing before starting REPL',
-          programmaticTest,
-          {
-            evalCodeBefore: `${setReportGlobal('repl')};${saveReportsAsGlobal}`,
-            stdinCode: '',
-          },
-          (stdout) => {
-            exp(globalInRepl.testReport).toMatchObject({
-              stdinReport: false,
-              evalReport: false,
-              replReport: {
-                __filename: false,
-                __dirname: false,
-
-                // Due to limitations in node's REPL API, we can't really expose
-                // the `module` prior to calling repl.start() which also sends
-                // output to stdout.
-                // For now, leaving this as unsupported / undefined behavior.
-
-                // moduleId: '<repl>',
-                // modulePath: '.',
-                // moduleFilename: null,
-                // modulePaths: [
-                //   join(ROOT_DIR, `repl/node_modules`),
-                //   ...rootModulePaths,
-                //   join(homedir(), `.node_modules`),
-                //   join(homedir(), `.node_libraries`),
-                //   // additional entry goes to node's install path
-                //   exp.any(String),
-                // ],
-                // // Note: vanilla node REPL does not set exports
-                // exportsTest: true,
-                // moduleAccessorsTest: true,
-
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(ROOT_DIR, '<repl>.ts')}:1:`
-                ),
-              },
-            });
-          }
-        );
-        test.serial(
-          'programmatically, passing code to stdin after starting REPL',
-          programmaticTest,
-          {
-            evalCodeBefore: null,
-            stdinCode: `${setReportGlobal('repl')};${saveReportsAsGlobal}`,
-          },
-          (stdout) => {
-            exp(globalInRepl.testReport).toMatchObject({
-              stdinReport: false,
-              evalReport: false,
-              replReport: {
-                __filename: false,
-                __dirname: false,
-                moduleId: '<repl>',
-                modulePath: '.',
-                moduleFilename: null,
-                modulePaths: exp.objectContaining({
-                  ...[join(ROOT_DIR, `repl/node_modules`), ...rootModulePaths],
-                }),
-                // Note: vanilla node REPL does not set exports
-                exportsTest: true,
-                // Note: vanilla node uses different name. See #1360
-                stackTest: exp.stringContaining(
-                  `    at ${join(ROOT_DIR, '<repl>.ts')}:1:`
-                ),
-                moduleAccessorsTest: true,
-              },
-            });
-            // Prior to these, nyc adds another entry on Windows; we need to ignore it
-            exp(
-              globalInRepl.testReport.replReport.modulePaths.slice(-3)
-            ).toMatchObject([
-              join(homedir(), `.node_modules`),
-              join(homedir(), `.node_libraries`),
-              // additional entry goes to node's install path
-              exp.any(String),
-            ]);
-          }
-        );
-      }
-    );
-
-    test.suite(
-      'REPL ignores diagnostics that are annoying in interactive sessions',
-      (test) => {
-        const code = `function foo() {};\nfunction foo() {return 123};\nconsole.log(foo());\n`;
-        const diagnosticMessage = `Duplicate function implementation`;
-        test(
-          'interactive repl should ignore them',
-          execMacro,
-          {
-            flags: '-i',
-            stdin: code,
-          },
-          async (stdout, stderr) => {
-            exp(stdout).not.toContain(diagnosticMessage);
-          }
-        );
-        test(
-          'interactive repl should not ignore them if they occur in other files',
-          execMacro,
-          {
-            flags: '-i',
-            stdin: `import './repl-ignored-diagnostics/index.ts';\n`,
-          },
-          async (stdout, stderr) => {
-            exp(stderr).toContain(diagnosticMessage);
-          }
-        );
-        test(
-          '[stdin] should not ignore them',
-          execMacro,
-          {
-            stdin: code,
-            expectError: true,
-          },
-          async (stdout, stderr) => {
-            exp(stderr).toContain(diagnosticMessage);
-          }
-        );
-        test(
-          '[eval] should not ignore them',
-          execMacro,
-          {
-            flags: `-e "${code.replace(/\n/g, '')}"`,
-            expectError: true,
-          },
-          async (stdout, stderr) => {
-            exp(stderr).toContain(diagnosticMessage);
-          }
-        );
-      }
-    );
 
     test('should support require flags', async () => {
       const { err, stdout } = await exec(
@@ -1884,141 +1188,6 @@ test.suite('ts-node', (test) => {
         });
         expect(err).to.equal(null);
         expect(stdout).to.contain('CommonJS');
-      });
-    }
-  });
-
-  test.suite('top level await', (test) => {
-    const compilerOptions = {
-      target: 'es2018',
-    };
-    function executeInTlaRepl(input: string, waitMs = 1000) {
-      return executeInRepl(
-        input
-          .split('\n')
-          .map((line) => line.trim())
-          // Restore newline once https://github.com/nodejs/node/pull/39392 is merged
-          .join(''),
-        {
-          waitMs,
-          createServiceOpts: {
-            experimentalReplAwait: true,
-            compilerOptions,
-          },
-          startOptions: { useGlobal: false },
-        }
-      );
-    }
-
-    if (semver.gte(ts.version, '3.8.0')) {
-      // Serial because it's timing-sensitive
-      test.serial('should allow evaluating top level await', async () => {
-        const script = `
-        const x: number = await new Promise((r) => r(1));
-        for await (const x of [1,2,3]) { console.log(x) };
-        for (const x of ['a', 'b']) { await x; console.log(x) };
-        class Foo {}; await 1;
-        function Bar() {}; await 2;
-        const {y} = await ({y: 2});
-        const [z] = await [3];
-        x + y + z;
-      `;
-
-        const { stdout, stderr } = await executeInTlaRepl(script);
-        expect(stderr).to.equal('');
-        expect(stdout).to.equal('> 1\n2\n3\na\nb\n6\n> ');
-      });
-
-      // Serial because it's timing-sensitive
-      test.serial(
-        'should wait until promise is settled when awaiting at top level',
-        async () => {
-          const awaitMs = 500;
-          const script = `
-          const startTime = new Date().getTime();
-          await new Promise((r) => setTimeout(() => r(1), ${awaitMs}));
-          const endTime = new Date().getTime();
-          endTime - startTime;
-        `;
-          const { stdout, stderr } = await executeInTlaRepl(script, 6000);
-
-          expect(stderr).to.equal('');
-
-          const elapsedTime = Number(
-            stdout.split('\n')[0].replace('> ', '').trim()
-          );
-          expect(elapsedTime).to.be.gte(awaitMs - 50);
-          expect(elapsedTime).to.be.lte(awaitMs + 100);
-        }
-      );
-
-      // Serial because it's timing-sensitive
-      test.serial(
-        'should not wait until promise is settled when not using await at top level',
-        async () => {
-          const script = `
-          const startTime = new Date().getTime();
-          (async () => await new Promise((r) => setTimeout(() => r(1), ${1000})))();
-          const endTime = new Date().getTime();
-          endTime - startTime;
-        `;
-          const { stdout, stderr } = await executeInTlaRepl(script);
-
-          expect(stderr).to.equal('');
-
-          const ellapsedTime = Number(
-            stdout.split('\n')[0].replace('> ', '').trim()
-          );
-          expect(ellapsedTime).to.be.gte(0);
-          expect(ellapsedTime).to.be.lte(10);
-        }
-      );
-
-      // Serial because it's timing-sensitive
-      test.serial(
-        'should error with typing information when awaited result has type mismatch',
-        async () => {
-          const { stdout, stderr } = await executeInTlaRepl(
-            'const x: string = await 1'
-          );
-
-          expect(stdout).to.equal('> > ');
-          expect(stderr.replace(/\r\n/g, '\n')).to.equal(
-            '<repl>.ts(2,7): error TS2322: ' +
-              (semver.gte(ts.version, '4.0.0')
-                ? `Type 'number' is not assignable to type 'string'.\n`
-                : `Type '1' is not assignable to type 'string'.\n`) +
-              '\n'
-          );
-        }
-      );
-
-      // Serial because it's timing-sensitive
-      test.serial(
-        'should error with typing information when importing a file with type errors',
-        async () => {
-          const { stdout, stderr } = await executeInTlaRepl(
-            `const {foo} = await import('./tests/repl/tla-import');`
-          );
-
-          expect(stdout).to.equal('> > ');
-          expect(stderr.replace(/\r\n/g, '\n')).to.equal(
-            'tests/repl/tla-import.ts(1,14): error TS2322: ' +
-              (semver.gte(ts.version, '4.0.0')
-                ? `Type 'number' is not assignable to type 'string'.\n`
-                : `Type '1' is not assignable to type 'string'.\n`) +
-              '\n'
-          );
-        }
-      );
-
-      test('should pass upstream test cases', async () =>
-        upstreamTopLevelAwaitTests({ TEST_DIR, create, createRepl }));
-    } else {
-      test('should throw error when attempting to use top level await on TS < 3.8', async () => {
-        exp(executeInTlaRepl('', 1000)).rejects.toThrow(
-          'Experimental REPL await is not compatible with TypeScript versions older than 3.8'
-        );
       });
     }
   });
